+++ shortcuts-raycast-extension/src/model/internal/subtitle-formatter.ts
